+++ src/main/java/com/y4kstudios/pytestimp/fixtures/LambdaFixture.kt
@@ -56,6 +56,8 @@ internal fun getFixture(element: PyNamedParameter, typeEvalContext: TypeEvalCont
     val func = PsiTreeUtil.getParentOfType(element, PyCallable::class.java) ?: return null
     return getFixture(element.name, func, typeEvalContext)
 }
+fun PyNamedParameter.isFixture(typeEvalContext: TypeEvalContext) = getFixture(this, typeEvalContext) != null
+
 
 internal fun getFixture(name: String?, source: PsiElement, typeEvalContext: TypeEvalContext): PyTestFixture? {
     return getFixtures(source, typeEvalContext).firstOrNull { o -> o.name == name }